backend/requirements.txt
waitress
pandas
flask
scikit-learn
tensorflow-cpu
numpy